refactor: Graph Processing
Refactored the graph processing out to be less tied to specific flows so that I can embed a graph flow as part of the main process as a means of allowing for dynamic identification of files that should be processed. Also introduced a special case node that can be used to export runtime variables from one graph to be used in subsequent graphs so that the pre-process can setup a working directory, the main process can use it and the post-process can clear it. Also can be used for passing out the collection of taret files that need to be processed