add: Nodes for Manifest Data and Comparison
Setup nodes that will allow for the storage of generic manifest data in a json format. This is so that we can store state values for the pipeline that can be used to influence behaviour on subsequent runs and allow for the skipping of files that have already been processed. Comparison node is included to allow us to test the state values that are required for processing